New and updated version of SWORD and ObjCSword wrapper. Improvements in module manager. Refreshing install source does not require to re-select the install source to show content. Memory management improvements. There is still a problem after editing an install source. Working on that.

+++ Resources/ObjCSword.framework/Versions/A/Headers/SwordVerseKey.h
@@ -33,7 +33,7 @@
 - (SwordVerseKey *)initWithRef:(NSString *)aRef;
 - (SwordVerseKey *)initWithRef:(NSString *)aRef v11n:(NSString *)scheme;
 
-- (int)index;
+- (long)index;
 - (int)testament;
 - (void)setTestament:(char)val;
 - (int)book;
